Fix Rack adapter and ensure the headers are sent in the response
