+++ examples/rack_server.rb
@@ -4,7 +4,7 @@ require 'bundler/setup'
 require 'polyphony/http'
 
 app_path = ARGV.first || File.expand_path('./config.ru', __dir__)
-app = Polyphony::HTTP::Rack.load(app_path)
+app = Polyphony::HTTP::Server::RackAdapter.load(app_path)
 opts = { reuse_addr: true, dont_linger: true }
 
 puts 'listening on port 1234'
