+++ lib/polyphony/http/server/http2_adapter.rb
